Colocate gql query
